Spanish translation of p03_ch02_the_null_zone_revisited.md
Greetings!

Another translation here. I got aware of the note about "reverting" the past Spanish translation, and since the string was not updated to Spanish I made the translation from scratch. I also saw ways to improve it (in grammar, contextual meaning and redaction).

I hope this helps to the whole translation of the book!

///// Alejandro /////